Add RSA and AES-CBC encryption

+++ include/OneControl/Constants.h
@@ -8,7 +8,7 @@ namespace oc
 	/**
 	 * Used for specifying which version of the software is being used to connected machines.
 	 */
-	using VersionInt = uint32_t; // Versioning scheme type
+	using VersionInt = uint16_t; // Versioning scheme type
 
 	/**
 	 * The default port on which the application will be running.
